# functional-programming/src/main/java/com/springboot/project/transaction/model/TransactionFilterRequestModel.java
package com.springboot.project.transaction.model;

import com.springboot.project.transaction.generated.dto.DomainEnumDto;
import com.springboot.project.common.generated.dto.PaginationRequestDto;
import com.springboot.project.transaction.generated.dto.PaymentMethodEnumDto;
import com.springboot.project.transaction.generated.dto.TransactionStatusEnumDto;
import java.math.BigDecimal;
import java.time.LocalDate;
import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class TransactionFilterRequestModel {

    private LocalDate date;

    private DomainEnumDto domain;

    private String location;

    private BigDecimal value;

    private TransactionStatusEnumDto status;

    private PaymentMethodEnumDto paymentMethod;

    private BigDecimal taxAmount;

    private BigDecimal netValue;

    private PaginationRequestDto pagination;
}